requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-pubsub
google-cloud-functions
requests
python-dotenv
